requirements.txt
pandas
numpy
json
matplotlib
python-binance
sqlalchemy
dash
dash-core-components
dash-html-components